Promote Azure Native to be the default Azure provider
