+++ azure-classic-csharp/${PROJECT}.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Pulumi.AzureNextGen" Version="0.*" />
+    <PackageReference Include="Pulumi.Azure" Version="3.*" />
   </ItemGroup>
 
 </Project>
