+++ azure-classic-csharp/.gitignore
