+++ azure-classic-fsharp/.gitignore
