+++ azure-classic-go/Pulumi.yaml
@@ -1,8 +1,8 @@
 name: ${PROJECT}
 description: ${DESCRIPTION}
-runtime: dotnet
+runtime: go
 template:
-  description: "A minimal Azure VB.NET Pulumi program"
+  description: A minimal Go Pulumi program with the classic Azure provider
   config:
     azure:location:
       description: The Azure location to use
