+++ azure-classic-go/go.mod
@@ -3,6 +3,6 @@ module ${PROJECT}
 go 1.14
 
 require (
-	github.com/pulumi/pulumi-azure-nextgen/sdk v0.5.1
+	github.com/pulumi/pulumi-azure/sdk/v3 v3.45.0
 	github.com/pulumi/pulumi/sdk/v2 v2.19.0
 )
