+++ azure-classic-javascript/.gitignore
