+++ azure-classic-javascript/index.js
