+++ azure-classic-javascript/package.json
