+++ azure-classic-python/.gitignore
