+++ azure-classic-python/requirements.txt
@@ -1,0 +1,2 @@
+pulumi>=2.0.0,<3.0.0
+pulumi-azure>=3.0.0,<4.0.0
