+++ azure-classic-typescript/.gitignore
