+++ azure-classic-typescript/package.json
@@ -5,6 +5,6 @@
     },
     "dependencies": {
         "@pulumi/pulumi": "^2.0.0",
-        "@pulumi/azure-nextgen": "^0.3.0"
+        "@pulumi/azure": "^3.0.0"
     }
 }
