+++ azure-classic-typescript/tsconfig.json
