+++ azure-classic-visualbasic/.gitignore
