+++ azure-classic-visualbasic/MyStack.vb
