+++ azure-csharp/${PROJECT}.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Pulumi.Azure" Version="3.*" />
+    <PackageReference Include="Pulumi.AzureNative" Version="0.*" />
   </ItemGroup>
 
 </Project>
