+++ azure-csharp/MyStack.cs
@@ -1,6 +1,8 @@
+using System.Threading.Tasks;
 using Pulumi;
-using Pulumi.Azure.Core;
-using Pulumi.Azure.Storage;
+using Pulumi.AzureNative.Resources.Latest;
+using Pulumi.AzureNative.Storage.Latest;
+using Pulumi.AzureNative.Storage.Latest.Inputs;
 
 class MyStack : Stack
 {
@@ -9,18 +11,32 @@ class MyStack : Stack
         // Create an Azure Resource Group
         var resourceGroup = new ResourceGroup("resourceGroup");
 
-        // Create an Azure Storage Account
-        var storageAccount = new Account("storage", new AccountArgs
+        // Create an Azure resource (Storage Account)
+        var storageAccount = new StorageAccount("sa", new StorageAccountArgs
         {
             ResourceGroupName = resourceGroup.Name,
-            AccountReplicationType = "LRS",
-            AccountTier = "Standard"
+            Sku = new SkuArgs
+            {
+                Name = SkuName.Standard_LRS
+            },
+            Kind = Kind.StorageV2
         });
 
-        // Export the connection string for the storage account
-        this.ConnectionString = storageAccount.PrimaryConnectionString;
+        // Export the primary key of the Storage Account
+        this.PrimaryStorageKey = Output.Tuple(resourceGroup.Name, storageAccount.Name).Apply(names =>
+            Output.CreateSecret(GetStorageAccountPrimaryKey(names.Item1, names.Item2)));
     }
 
     [Output]
-    public Output<string> ConnectionString { get; set; }
+    public Output<string> PrimaryStorageKey { get; set; }
+
+    private static async Task<string> GetStorageAccountPrimaryKey(string resourceGroupName, string accountName)
+    {
+        var accountKeys = await ListStorageAccountKeys.InvokeAsync(new ListStorageAccountKeysArgs
+        {
+            ResourceGroupName = resourceGroupName,
+            AccountName = accountName
+        });
+        return accountKeys.Keys[0].Value;
+    }
 }
