+++ azure-fsharp/Program.fs
@@ -1,23 +1,39 @@
 ﻿module Program
 
 open Pulumi.FSharp
-open Pulumi.Azure.Core
-open Pulumi.Azure.Storage
+open Pulumi.AzureNative.Resources.Latest
+open Pulumi.AzureNative.Storage.Latest
+open Pulumi.AzureNative.Storage.Latest.Inputs
+
+// Helper function to retrieve the primary key of a storage account
+let getStorageAccountPrimaryKey(resourceGroupName: string, accountName: string): Async<string> = async {
+    let! accountKeys =
+        ListStorageAccountKeysArgs(ResourceGroupName = resourceGroupName, AccountName = accountName) |>
+        ListStorageAccountKeys.InvokeAsync
+        |> Async.AwaitTask
+    return accountKeys.Keys.[0].Value
+}
 
 let infra () =
     // Create an Azure Resource Group
-    let resourceGroup = ResourceGroup "resourceGroup"
+    let resourceGroup =
+        ResourceGroup("resourceGroup")
 
     // Create an Azure Storage Account
     let storageAccount =
-        Account("storage",
-            AccountArgs
-               (ResourceGroupName = io resourceGroup.Name,
-                AccountReplicationType = input "LRS",
-                AccountTier = input "Standard"))
-
-    // Export the connection string for the storage account
-    dict [("connectionString", storageAccount.PrimaryConnectionString :> obj)]
+        StorageAccount("sa", 
+            StorageAccountArgs
+                (ResourceGroupName = io resourceGroup.Name,
+                 Sku = input (SkuArgs(Name = inputUnion2Of2 SkuName.Standard_LRS)),
+                 Kind = inputUnion2Of2 Kind.StorageV2))
+        
+    // Get the primary key
+    let primaryKey =
+        Outputs.pair resourceGroup.Name storageAccount.Name
+        |> Outputs.applyAsync getStorageAccountPrimaryKey
+    
+    // Export the primary key for the storage account
+    dict [("connectionString", primaryKey :> obj)]
 
 [<EntryPoint>]
 let main _ =
