--- azure-javascript/Pulumi.yaml
@@ -1,13 +1,0 @@
-name: ${PROJECT}
-description: ${DESCRIPTION}
-runtime: nodejs
-template:
-  description: A minimal Azure JavaScript Pulumi program
-  important: true
-  config:
-    azure:environment:
-      description: The Azure environment to use (`public`, `usgovernment`, `german`, `china`)
-      default: public
-    azure:location:
-      description: The Azure location to use
-      default: WestUS
