--- azure-nextgen-csharp/MyStack.cs
@@ -1,48 +1,0 @@
-using System.Threading.Tasks;
-using Pulumi;
-using Pulumi.AzureNextGen.Resources.Latest;
-using Pulumi.AzureNextGen.Storage.Latest;
-using Pulumi.AzureNextGen.Storage.Latest.Inputs;
-
-class MyStack : Stack
-{
-    public MyStack()
-    {
-        // Create an Azure Resource Group
-        var resourceGroup = new ResourceGroup("resourceGroup", new ResourceGroupArgs
-        {
-            ResourceGroupName = "my-rg",
-            Location = "WestUS"
-        });
-
-        // Create an Azure resource (Storage Account)
-        var storageAccount = new StorageAccount("sa", new StorageAccountArgs
-        {
-            ResourceGroupName = resourceGroup.Name,
-            AccountName = "mystorageaccount",
-            Location = resourceGroup.Location,
-            Sku = new SkuArgs
-            {
-                Name = SkuName.Standard_LRS
-            },
-            Kind = Kind.StorageV2
-        });
-
-        // Export the primary key of the Storage Account
-        this.PrimaryStorageKey = Output.Tuple(resourceGroup.Name, storageAccount.Name).Apply(names =>
-            Output.CreateSecret(GetStorageAccountPrimaryKey(names.Item1, names.Item2)));
-    }
-
-    [Output]
-    public Output<string> PrimaryStorageKey { get; set; }
-
-    private static async Task<string> GetStorageAccountPrimaryKey(string resourceGroupName, string accountName)
-    {
-        var accountKeys = await ListStorageAccountKeys.InvokeAsync(new ListStorageAccountKeysArgs
-        {
-            ResourceGroupName = resourceGroupName,
-            AccountName = accountName
-        });
-        return accountKeys.Keys[0].Value;
-    }
-}
