--- azure-nextgen-csharp/Pulumi.yaml
@@ -1,5 +1,0 @@
-name: ${PROJECT}
-description: ${DESCRIPTION}
-runtime: dotnet
-template:
-  description: "A minimal Azure NextGen C# Pulumi program"
