--- azure-nextgen-go/Pulumi.yaml
@@ -1,5 +1,0 @@
-name: ${PROJECT}
-description: ${DESCRIPTION}
-runtime: go
-template:
-  description: A minimal Azure NextGen Go Pulumi program
