--- azure-nextgen-go/main.go
@@ -1,54 +1,0 @@
-package main
-
-import (
-	resources "github.com/pulumi/pulumi-azure-nextgen/sdk/go/azure/resources/latest"
-	storage "github.com/pulumi/pulumi-azure-nextgen/sdk/go/azure/storage/latest"
-	"github.com/pulumi/pulumi/sdk/v2/go/pulumi"
-)
-
-func main() {
-	pulumi.Run(func(ctx *pulumi.Context) error {
-		// Create an Azure Resource Group
-		resourceGroup, err := resources.NewResourceGroup(ctx, "resourceGroup", &resources.ResourceGroupArgs{
-			ResourceGroupName: pulumi.String("my-rg"),
-			Location:          pulumi.String("WestUS"),
-		})
-		if err != nil {
-			return err
-		}
-
-		// Create an Azure resource (Storage Account)
-		account, err := storage.NewStorageAccount(ctx, "sa", &storage.StorageAccountArgs{
-			ResourceGroupName: resourceGroup.Name,
-			AccountName:       pulumi.String("mystorageaccount"),
-			Location:          resourceGroup.Location,
-			AccessTier:        storage.AccessTierHot,
-			Sku: &storage.SkuArgs{
-				Name: storage.SkuName_Standard_LRS,
-			},
-			Kind: storage.KindStorageV2,
-		})
-		if err != nil {
-			return err
-		}
-
-		// Export the primary key of the Storage Account
-		ctx.Export("primaryStorageKey", pulumi.All(resourceGroup.Name, account.Name).ApplyT(
-			func(args []interface{}) (string, error) {
-				resourceGroupName := args[0].(string)
-				accountName := args[1].(string)
-				accountKeys, err := storage.ListStorageAccountKeys(ctx, &storage.ListStorageAccountKeysArgs{
-					ResourceGroupName: resourceGroupName,
-					AccountName:       accountName,
-				})
-				if err != nil {
-					return "", err
-				}
-
-				return accountKeys.Keys[0].Value, nil
-			},
-		))
-
-		return nil
-	})
-}
