--- azure-nextgen-python/Pulumi.yaml
@@ -1,5 +1,0 @@
-name: ${PROJECT}
-description: ${DESCRIPTION}
-runtime: python
-template:
-  description: A minimal Azure NextGen Python Pulumi program
