--- azure-nextgen-python/requirements.txt
@@ -1,2 +1,0 @@
-pulumi>=2.0.0,<3.0.0
-pulumi-azure-nextgen>=0.3.0
