--- azure-nextgen-typescript/Pulumi.yaml
@@ -1,5 +1,0 @@
-name: ${PROJECT}
-description: ${DESCRIPTION}
-runtime: nodejs
-template:
-  description: A minimal Azure NextGen TypeScript Pulumi program
