--- azure-nextgen-typescript/index.ts
@@ -1,25 +1,0 @@
-import * as pulumi from "@pulumi/pulumi";
-import * as resources from "@pulumi/azure-nextgen/resources/latest";
-import * as storage from "@pulumi/azure-nextgen/storage/latest";
-
-// Create an Azure Resource Group
-const resourceGroup = new resources.ResourceGroup("resourceGroup", {
-    resourceGroupName: "my-rg",
-    location: "westus",
-});
-
-// Create an Azure resource (Storage Account)
-const storageAccount = new storage.StorageAccount("sa", {
-    resourceGroupName: resourceGroup.name,
-    accountName: "mystorageaccount",
-    location: resourceGroup.location,
-    sku: {
-        name: storage.SkuName.Standard_LRS,
-    },
-    kind: storage.Kind.StorageV2,
-});
-
-// Export the primary key of the Storage Account
-const storageAccountKeys = pulumi.all([resourceGroup.name, storageAccount.name]).apply(([resourceGroupName, accountName]) =>
-    storage.listStorageAccountKeys({ resourceGroupName, accountName }));
-export const primaryStorageKey = storageAccountKeys.keys[0].value;
