+++ azure-python/__main__.py
@@ -1,17 +1,25 @@
-"""An Azure Python Pulumi program"""
+"""An Azure RM Python Pulumi program"""
 
 import pulumi
-from pulumi_azure import core, storage
+from pulumi_azure_native.storage import latest as storage
+from pulumi_azure_native.resources import latest as resources
 
 # Create an Azure Resource Group
-resource_group = core.ResourceGroup('resource_group')
+resource_group = resources.ResourceGroup('resource_group')
 
 # Create an Azure resource (Storage Account)
-account = storage.Account('storage',
-                          # The location for the storage account will be derived automatically from the resource group.
-                          resource_group_name=resource_group.name,
-                          account_tier='Standard',
-                          account_replication_type='LRS')
+account = storage.StorageAccount('sa',
+    resource_group_name=resource_group.name,
+    sku=storage.SkuArgs(
+        name=storage.SkuName.STANDARD_LRS,
+    ),
+    kind=storage.Kind.STORAGE_V2)
 
-# Export the connection string for the storage account
-pulumi.export('connection_string', account.primary_connection_string)
+# Export the primary key of the Storage Account
+primary_key = pulumi.Output.all(resource_group.name, account.name) \
+    .apply(lambda args: storage.list_storage_account_keys(
+        resource_group_name=args[0],
+        account_name=args[1]
+    )).apply(lambda accountKeys: accountKeys.keys[0].value)
+
+pulumi.export("primary_storage_key", primary_key)
