+++ azure-python/requirements.txt
@@ -1,2 +1,2 @@
 pulumi>=2.0.0,<3.0.0
-pulumi-azure>=3.0.0,<4.0.0
+pulumi-azure-native>=0.7.0
