+++ azure-typescript/package.json
@@ -5,6 +5,6 @@
     },
     "dependencies": {
         "@pulumi/pulumi": "^2.0.0",
-        "@pulumi/azure": "^3.0.0"
+        "@pulumi/azure-native": "^0.7.0"
     }
 }
